requirements.txt
discord.py
python-dotenv
duckduckgo-search
google-genai
google-generativeai
deep_translator
